requirements.txt
Flask==3.0.2
flasgger==0.9.7.1
python-dotenv==1.0.1
Flask-Cors==4.0.0
flask-restx==1.3.0
Flask-Injector 
pyspark 
pyspark[sql]
pyspark[pandas_on_spark]
plotly
pyspark[connect]
pandas
numpy
matplotlib
pyarrow
reportlab==4.1.0
langchain-openai
langchain